code/requirements_sheets.txt
# Function dependencies, for example:
# package>=version
gspread
pandas-gbq
pandas
google-auth